sidecar/requirements.txt
pyicloud-ipd==0.10.0
python-dateutil==2.9.0.post0
pyinstaller==6.12.0
